refactor: extract mongodb wrapper from mongo.integration.base.tests.js
